requirements.txt
requests==2.22.0
Django==1.10.5
PyGithub==1.43.8
